Fix -h option description
Currently if you run moss with no arguments the output is as follows:
/bin/moss: [command] [--options]

Options:
   --version Show the program version and exit
-D --destdir Set the system root
-h    --help This help information.

This is incorrect as adding the option -h will change the output so "this" can't
be the help information. This commit changes the description to the better fitting
"Show help information" which properly describes the option.

Signed-off-by: Sam Smith <sam.henning.smith@protonmail.com>